core: split tee_ta_manager.c
Isolates static and user TA specifics into separate files. Almost no change
in behaviour, except some changes for static TA to be more consistent.
