+++ core/arch/arm/kernel/sub.mk
@@ -1,4 +1,6 @@
 srcs-y += tee_ta_manager.c
+srcs-y += user_ta.c
+srcs-y += static_ta.c
 srcs-y += elf_load.c
 srcs-y += tee_time.c
